encode tf record fitting train, valid, and test

+++ modeling/vectorMaker.py
@@ -149,41 +149,43 @@ class VectorMaker:
             self.vector_matrix[key] = matrix
 
     def build_tf_records(self):
-        if os.path.isdir(self.tf_record_path):
-            print("\nThe directory for tfrecord is already existed -", self.tf_record_path, "\n")
-            while True:
-                do_continue = input("Do you want to re-encoding? (y/n) - ").lower()
-                if do_continue == 'n':
-                    return
-                elif do_continue == 'y':
-                    shutil.rmtree(self.tf_record_path)
-                    break
+        if VERSION == 1:
+            if os.path.isdir(self.tf_record_path):
+                print("\nThe directory for tfrecord is already existed -", self.tf_record_path, "\n")
+                while True:
+                    do_continue = input("Do you want to re-encoding? (y/n) - ").lower()
+                    if do_continue == 'n':
+                        return
+                    elif do_continue == 'y':
+                        shutil.rmtree(self.tf_record_path)
+                        break
 
-        os.mkdir(self.tf_record_path)
-        x_train, y_train = self.__get_set(key="train")
-        x_valid, y_valid = self.__get_set(key="valid")
-        x_test, y_test = self.__get_set(key="test")
-        x_data, y_data = self.__get_shuffle_set(x_train + x_valid + x_test, y_train + y_valid + y_test)
+            os.mkdir(self.tf_record_path)
+            x_train, y_train = self.__get_set(key="train")
+            x_valid, y_valid = self.__get_set(key="valid")
+            x_test, y_test = self.__get_set(key="test")
+            x_data, y_data = self.__get_shuffle_set(x_train + x_valid + x_test, y_train + y_valid + y_test)
 
-        # shuffle data for avoiding over-fitting
-        if IS_CROSS_VALID:
-            print("This scope will be implemented")
-            # TODO implement k-fold cross validation
-            exit(-1)
-            for x_train, y_train, x_test, y_test in self.__data_generator(x_data, y_data):
+            # shuffle data for avoiding over-fitting
+            if IS_CROSS_VALID:
+                print("This scope will be implemented")
+                # TODO implement k-fold cross validation
+                exit(-1)
+                for x_train, y_train, x_test, y_test in self.__data_generator(x_data, y_data):
+                    self.tf_recorder.to_tf_records(x_train, y_train, key="train")
+                    self.tf_recorder.to_tf_records(x_test, y_test, key="test")
+            else:
+                if DO_ENCODE_IMAGE:
+                    i_train, i_valid = int(len(y_data) * TRAIN_RATIO), int(len(y_data) * VALID_RATIO)
+                    x_train, x_valid, x_test = x_data[:i_train], x_data[i_train:i_valid], x_data[i_valid:]
+                    y_train, y_valid, y_test = y_data[:i_train], y_data[i_train:i_valid], y_data[i_valid:]
+
                 self.tf_recorder.to_tf_records(x_train, y_train, key="train")
+                self.tf_recorder.to_tf_records(x_valid, y_valid, key="valid")
                 self.tf_recorder.to_tf_records(x_test, y_test, key="test")
-        else:
-            i_train, i_valid = int(len(y_data) * TRAIN_RATIO), int(len(y_data) * VALID_RATIO)
-            x_train, x_valid, x_test = x_data[:i_train], x_data[i_train:i_valid], x_data[i_valid:]
-            y_train, y_valid, y_test = y_data[:i_train], y_data[i_train:i_valid], y_data[i_valid:]
 
-            self.tf_recorder.to_tf_records(x_train, y_train, key="train")
-            self.tf_recorder.to_tf_records(x_valid, y_valid, key="valid")
-            self.tf_recorder.to_tf_records(x_test, y_test, key="test")
-
-        self.tf_recorder.save()
-        print("success build tf records! (in the -", self.tf_record_path + ")\n\n\n")
+            self.tf_recorder.save()
+            print("success build tf records! (in the -", self.tf_record_path + ")\n\n\n")
 
     def __add_key_value_in_dict(self, key, value):
         if KEY_TF_NAME not in self.vector_matrix:
